requirements.txt
pandas
numpy
alpaca-py
python-dotenv
pandas_ta
arch
statsmodels
skforecast
